created schedule route
