+++ routes/api/index.js
@@ -6,6 +6,9 @@ const router = require("express").Router();
 const apiRoutes = require("./apiRoutes");
 router.use("/", apiRoutes);
 
+
+
+
 // For anything else, render the html page
 router.use(function(req, res) {
   // TODO: delete this if not used
